requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cors==1.0.1
Django==5.1.6
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.17.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
setuptools==75.8.0
sqlparse==0.5.3
tldextract==5.1.3
tzdata==2025.1
urllib3==2.3.0
zope.event==5.0
zope.interface==7.2
coreapi===2.3.3